Fixes for grammar tables in current HTML
This change fixes the output for nouns and verbs in the current Duden
HTML.

A new strategy will be needed, though, for some of the grammar
tables on pages for adjectives (see e.g.
https://www.duden.de/rechtschreibung/leicht_zart_unkompliziert_geringfuegig#grammatik),
which contain additional information (e.g. articles) that doesn't
fit well into the current representation of table cells as tagged
strings.